LayoutTests/http/wpt/webrtc/video-script-transform-simulcast.html

<!doctype html>
<html>
    <head>
        <meta charset="utf-8">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="/webrtc/third_party/sdp/sdp.js"></script>
        <script src="/webrtc/simulcast/simulcast.js"></script>
    </head>
    <body>
        <video id="videoL" controls autoplay playsInline></video>
        <video id="videoM" controls autoplay playsInline></video>
        <script>
function waitForMessage(port, data)
{
    let gotMessage;
    const promise = new Promise((resolve, reject) => {
        gotMessage = resolve;
        setTimeout(() => { reject("did not get " + data) }, 5000);
    });
    port.onmessage = event => {
       if (event.data === data)
           gotMessage();
    };
    return promise;
}

let worker;
promise_test(async (test) => {
    worker = new Worker('context-transform.js');
    const data = await new Promise(resolve => worker.onmessage = (event) => resolve(event.data));
    assert_equals(data, "registered");

    const localStream = await navigator.mediaDevices.getUserMedia({video: true});

    const senderChannel = new MessageChannel;
    const receiverChannelL = new MessageChannel;
    const receiverChannelM = new MessageChannel;
    let sender, receiverM, receiverL;
    const senderTransform = new RTCRtpScriptTransform(worker, {name:'MockRTCRtpTransform', mediaType:'video', side:'sender', port:senderChannel.port2}, [senderChannel.port2]);
    const receiverTransformL = new RTCRtpScriptTransform(worker, {name:'MockRTCRtpTransform', mediaType:'video', side:'receiver', port:receiverChannelL.port2}, [receiverChannelL.port2]);
    const receiverTransformM = new RTCRtpScriptTransform(worker, {name:'MockRTCRtpTransform', mediaType:'video', side:'receiver', port:receiverChannelM.port2}, [receiverChannelM.port2]);
    senderTransform.port = senderChannel.port1;
    receiverTransformL.port = receiverChannelL.port1;
    receiverTransformM.port = receiverChannelM.port1;

    const promise1 = waitForMessage(senderTransform.port, "started video sender");
    const promise2 = waitForMessage(receiverTransformL.port, "started video receiver");
    const promise3 = waitForMessage(receiverTransformM.port, "started video receiver");

    await Promise.all([promise1, promise2, promise3]);

    const pc1 = new RTCPeerConnection();
    const pc2 = new RTCPeerConnection();

    pc1.onicecandidate = e => {
        const candidate = e.candidate;
        if (!candidate)
            return;
        const newCandidate = { candidate: candidate.candidate, sdpMid: "l", sdpMLineIndex:candidate.sdpMLineIndex };
        pc2.addIceCandidate(newCandidate);
    }
    pc2.onicecandidate = e => {
        const candidate = e.candidate;
        if (!candidate)
            return;
        const newCandidate = { candidate: candidate.candidate, sdpMid: "0", sdpMLineIndex:candidate.sdpMLineIndex };
        pc1.addIceCandidate(newCandidate);
    }

    sender = pc1.addTransceiver("video", {sendEncodings: [{rid: "l", scaleResolutionDownBy: 2}, {rid: "m", scaleResolutionDownBy: 1}]}).sender;
    sender.replaceTrack(localStream.getVideoTracks()[0]);
    sender.setStreams(localStream);
    sender.transform = senderTransform;

    const streamPromise = new Promise((resolve, reject) => {
        pc2.ontrack = (trackEvent) => {
            if (!receiverL) {
                receiverL = trackEvent.receiver;
                receiverL.transform = receiverTransformL;
                videoL.srcObject = trackEvent.streams[0];
                return;
            }

            receiverM = trackEvent.receiver;
            receiverM.transform = receiverTransformM;
            videoM.srcObject = trackEvent.streams[0];
            resolve();
        };
        setTimeout(() => reject("Getting stream timed out"), 5000);
    });

    const offer = await pc1.createOffer();
    await pc1.setLocalDescription(offer);
    await pc2.setRemoteDescription({
        type: 'offer',
        sdp: swapRidAndMidExtensionsInSimulcastOffer(offer, ["l", "m"]),
    });
    const answer = await pc2.createAnswer();
    await pc2.setLocalDescription(answer);
    await pc1.setRemoteDescription({
        type: 'answer',
        sdp: swapRidAndMidExtensionsInSimulcastAnswer(answer, pc1.localDescription, ["l", "m"]),
    });

    await streamPromise;
}, "setup");

async function waitForVideoSize(video, width, height)
{
    const max = 200
    let counter = 0;
    while (++counter < max && video.videoWidth != width && video.videoHeight != height)
        await waitFor(50);

    if (counter === max)
        return Promise.reject("Video size not expected : " + video.videoWidth + " " + video.videoHeight);
}

promise_test(async () => {
    await Promise.all([videoL.play(), videoM.play()]);
    await waitForVideoSize(videoL, 320, 240);
    await waitForVideoSize(videoM, 640, 480);
}, "videos are playing with the correct resolutions");
        </script>
    </body>
</html>
